Made refactoring for admin.py
* Renamed admin_commands.py to admin.py
